somen/requirements.txt
mlflow
feather-format
tables
pytorch
ignite
scikit-learn
pandas
numpy
